requirements.txt
fastapi
uvicorn
pyspellchecker
